backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
ultralytics>=8.0.0
pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.21.0
